refac: move members to 'account.py'
