+++ src/finhan/main.py
@@ -1,35 +1,17 @@
 #!/usr/bin/env python3
-from datetime import timedelta
-from functools import reduce
-from itertools import chain, starmap
-from pathlib import Path
 from argparse import ArgumentParser
-from typing import Iterator, Iterable
+from pathlib import Path
 
-import numpy as np
 from matplotlib import pyplot
 from matplotlib.lines import Line2D
 
-import finhan.adapters.bepost as bepost
-from numpy.polynomial import polynomial
-
-from finhan.account import read_balance
+from finhan.account import read_balance, read_account_transactions, apply_balance, \
+    dates_and_numbers, joint_account_transactions
 
 
 def plot_grand_total(current_balance, transactions_by_account):
-    transactions_by_date = sorted(tuple(
-        chain(*transactions_by_account.values())
-    ), key=lambda t: t.date)
-
-    all_transactions = np.fromiter(
-        (t.amount for t in transactions_by_date),
-        dtype=float
-    )
-    floating_grand_total = np.cumsum(all_transactions)
-    last = floating_grand_total[-1]
-    grand_current_balance = sum(current_balance.values())
-    grand_total = floating_grand_total + (grand_current_balance - last)
-    dates = tuple(t.date for t in transactions_by_date)
+    dates, grand_total = joint_account_transactions(current_balance,
+                                                    transactions_by_account)
     pyplot.plot_date(dates, grand_total, '-', label=f'GRAND TOTAL')
 
 
@@ -47,67 +29,9 @@ def main():
     transactions_by_account = read_account_transactions(options.data_paths)
     plot_each_account(current_balance, transactions_by_account)
     plot_grand_total(current_balance, transactions_by_account)
+    pyplot.grid()
     pyplot.legend()
     pyplot.show()
-
-
-def read_account_transactions(data_paths):
-    line_lists = map(read_lines, data_paths)
-    transaction_lists = tuple(map(bepost.from_lines,
-                                  line_lists))
-    transaction_lists = tuple(
-        starmap(
-            lambda ts, account: (
-                tuple(ts), account),
-            transaction_lists
-        )
-    )
-    accounts = tuple(account for _, account in transaction_lists)
-
-    def for_account(account):
-        return (t for t, a in transaction_lists if a == account)
-
-    def join_transactions(t, other):
-        return tuple(t) + tuple(other)
-
-    transactions_by_account = {
-        account: reduce(join_transactions, for_account(account), tuple())
-        for account in accounts
-    }
-    return transactions_by_account
-
-
-def trend(dates, cumulative, label):
-    dates_u = unix_timestamps(dates)
-    p = polynomial.polyfit(dates_u, cumulative, 1)
-    line = np.poly1d(p)
-    edges, edges_u = find_edges(dates)
-    return (edges, line(edges_u)), f'{label}-trend: {line}'
-
-
-def find_edges(dates):
-    edges = (min(dates), max(dates) + timedelta(weeks=30))
-    edges_u = tuple(map(lambda d: d.timestamp(), edges))
-    return edges, edges_u
-
-
-def unix_timestamps(dates):
-    dates_t = np.empty(len(dates))
-    for i, d in enumerate(dates):
-        dates_t[i] = d.timestamp()
-    return dates_t
-
-
-def balance(current_balance, numbers):
-    saldo_floating = np.cumsum(numbers)
-    last_floating_saldo = saldo_floating[-1]
-    saldo = saldo_floating + (current_balance - last_floating_saldo)
-    return saldo
-
-
-def read_lines(filename: str) -> Iterator[str]:
-    with Path(filename).open() as f:
-        return f.readlines()
 
 
 def plot_each_account(current_balance, transactions_by_account):
@@ -120,20 +44,10 @@ def plot_each_account(current_balance, transactions_by_account):
 def plot_for_account(account, current_balance, transactions):
     dates, numbers = dates_and_numbers(transactions)
 
-    saldo = balance(current_balance, numbers)
+    balance = apply_balance(current_balance, numbers)
     line: Line2D = pyplot.plot_date(dates, numbers, label=account)[0]
-    pyplot.plot_date(dates, saldo, '-', color=line.get_color(),
-                     label=f'saldo {account}')
-
-
-def dates_and_numbers(transactions):
-    transactions = sorted(
-        transactions,
-        key=lambda t: t.date,
-        reverse=True)
-    dates = tuple(t.date for t in transactions)
-    numbers = tuple(t.amount for t in transactions)
-    return dates, numbers
+    pyplot.plot_date(dates, balance, '-x', color=line.get_color(),
+                     label=f'balance {account}')
 
 
 main()
